refactor: 공통 웹소켓 서버 훅 구현
