+++ src/pages/LearnSession.tsx
@@ -15,13 +15,9 @@ import StreamingControls from '@/components/StreamingControls';
 import { useMediaPipeHolistic } from '@/hooks/useMediaPipeHolistic';
 import { useBadgeSystem } from '@/hooks/useBadgeSystem';
 import { Button } from '@/components/ui/button';
+import { useClassifierClient } from '@/hooks/useClassifierClient';
 
-// 재시도 설정
-const RETRY_CONFIG = {
-  maxAttempts: 3,
-  initialDelay: 1000, // 1초
-  maxDelay: 5000, // 5초
-};
+
 
 const LearnSession = () => {
   const { categoryId, chapterId } = useParams();
@@ -30,38 +26,50 @@ const LearnSession = () => {
   const navigate = useNavigate();
   const location = useLocation();
   const [transmissionCount, setTransmissionCount] = useState(0);
-  const [lesson_mapper, setLessonMapper] = useState<{ [key: string]: string }>(location.state?.lesson_mapper || {});
-  const [currentWsUrl, setCurrentWsUrl] = useState<string>('');
-  const [currentConnectionId, setCurrentConnectionId] = useState<string>('');
-  const [retryAttempts, setRetryAttempts] = useState({
-    lessonMapper: 0,
-    wsConnection: 0,
-  });
-  const [isRetrying, setIsRetrying] = useState(false);
   const retryTimeoutRef = useRef<NodeJS.Timeout | null>(null);
-  const studyListRef = useRef<string[]>([]);
-  const [isBufferingPaused, setIsBufferingPaused] = useState(false);
-  const { connectionStatus, wsList, sendMessage } = useWebsocket();
 
-  const [displayConfidence, setDisplayConfidence] = useState<string>('');
+  // useClassifierClient 훅 사용
+  const {
+    isRetrying,
+    isConnected,
+    currentConnectionId,
+    currentWsUrl,
+    lessonMapper,
+    currentSignId,
+    currentSign,
+    currentResult,
+    feedback,
+    displayConfidence,
+    maxConfidence,
+    isBufferingPaused,
+    studyList,
+    setCurrentSignId,
+    setCurrentSign,
+    setLessonMapper,
+    setFeedback,
+    setDisplayConfidence,
+    setMaxConfidence,
+    setIsBufferingPaused,
+    retryLessonMapper,
+    retryWsConnection,
+    connectionStatus,
+    wsList,
+    sendMessage,
+  } = useClassifierClient();
 
-  const [isConnected, setIsConnected] = useState<boolean>(false); // 초기값에 의해 타입 결정됨.
-  const [currentResult, setCurrentResult] = useState<ClassificationResult | null>(null);
-  const [maxConfidence, setMaxConfidence] = useState(0.0);
-  const animationIntervalRef = useRef<NodeJS.Timeout | null>(null);
+  // 기존 상태들을 훅의 상태로 초기화
+  useEffect(() => {
+    setLessonMapper(location.state?.lesson_mapper || {});
+  }, [location.state?.lesson_mapper, setLessonMapper]);
 
   const { findCategoryById, findChapterById, findHierarchyByChapterId } = useLearningData();
 
-  const [animData, setAnimData] = useState(null);
-  const [currentFrame, setCurrentFrame] = useState(0);
-
   const [currentSignIndex, setCurrentSignIndex] = useState(0);
   const [lessons, setLessons] = useState<Lesson[]>([]);
-  const currentSign = lessons[currentSignIndex];
-  const currentSignId = lessons[currentSignIndex]?.id;
+  const currentLessonSign = lessons[currentSignIndex];
+  const currentLessonSignId = lessons[currentSignIndex]?.id;
   const [isRecording, setIsRecording] = useState(false);
 
-  const [feedback, setFeedback] = useState<'correct' | 'incorrect' | null>(null);
   const [sessionComplete, setSessionComplete] = useState(false);
 
   // 랜드마크 버퍼링 관련 상태
@@ -69,202 +77,13 @@ const LearnSession = () => {
   const bufferIntervalRef = useRef<NodeJS.Timeout | null>(null);
   const BUFFER_DURATION = 1000; // 2초
 
-  //===============================================
-  // 분류 서버 관련 훅
-  //===============================================
-
-  // lesson_mapper 재시도 함수
-  const retryLessonMapper = useCallback(async () => {
-    if (retryAttempts.lessonMapper >= RETRY_CONFIG.maxAttempts) {
-      console.error('[LearnSession] lesson_mapper 재시도 횟수 초과');
-      alert('데이터를 불러오는데 실패했습니다. 페이지를 새로고침하거나 다시 시도해주세요.');
-      setIsRetrying(false);
-      return;
-    }
-
-    setIsRetrying(true);
-    const delay = Math.min(
-      RETRY_CONFIG.initialDelay * Math.pow(2, retryAttempts.lessonMapper),
-      RETRY_CONFIG.maxDelay
-    );
-
-    console.log(`[LearnSession] lesson_mapper 재시도 ${retryAttempts.lessonMapper + 1}/${RETRY_CONFIG.maxAttempts} (${delay}ms 후)`);
-
-    retryTimeoutRef.current = setTimeout(() => {
-      // 이전 페이지로 돌아가서 다시 데이터 받아오기
-      if (location.state?.lesson_mapper && Object.keys(location.state.lesson_mapper).length > 0) {
-        setLessonMapper(location.state.lesson_mapper);
-        setRetryAttempts(prev => ({ ...prev, lessonMapper: 0 }));
-        // WebSocket 연결도 성공했거나 재시도가 필요없으면 전체 재시도 상태 해제
-        if (retryAttempts.wsConnection === 0 && currentConnectionId) {
-          setIsRetrying(false);
-        }
-        console.log('[LearnSession] lesson_mapper 재시도 성공');
-      } else {
-        setRetryAttempts(prev => ({ ...prev, lessonMapper: prev.lessonMapper + 1 }));
-        retryLessonMapper();
-      }
-    }, delay);
-  }, [retryAttempts.lessonMapper, retryAttempts.wsConnection, location.state, currentConnectionId]);
-
-  // WebSocket 연결 재시도 함수
-  const retryWsConnection = useCallback(async (targetUrl: string) => {
-    if (retryAttempts.wsConnection >= RETRY_CONFIG.maxAttempts) {
-      console.error('[LearnSession] WebSocket 연결 재시도 횟수 초과');
-      alert('서버 연결에 실패했습니다. 잠시 후 다시 시도해주세요.');
-      navigate("/");
-      setIsRetrying(false);
-      return;
-    }
-
-    setIsRetrying(true);
-    const delay = Math.min(
-      RETRY_CONFIG.initialDelay * Math.pow(2, retryAttempts.wsConnection),
-      RETRY_CONFIG.maxDelay
-    );
-
-    console.log(`[LearnSession] WebSocket 연결 재시도 ${retryAttempts.wsConnection + 1}/${RETRY_CONFIG.maxAttempts} (${delay}ms 후)`);
-
-    retryTimeoutRef.current = setTimeout(() => {
-      const connection = getConnectionByUrl(targetUrl);
-      if (connection) {
-        setCurrentConnectionId(connection.id);
-        setRetryAttempts(prev => ({ ...prev, wsConnection: 0 }));
-        // lesson_mapper도 성공했거나 재시도가 필요없으면 전체 재시도 상태 해제
-        if (retryAttempts.lessonMapper === 0 && Object.keys(lesson_mapper).length > 0) {
-          setIsRetrying(false);
-        }
-        console.log('[LearnSession] WebSocket 연결 재시도 성공:', connection.id);
-      } else {
-        setRetryAttempts(prev => ({ ...prev, wsConnection: prev.wsConnection + 1 }));
-        retryWsConnection(targetUrl);
-      }
-    }, delay);
-  }, [retryAttempts.wsConnection, retryAttempts.lessonMapper, lesson_mapper]);
-
-  // WebSocket 연결 상태 모니터링
+  // 현재 수어 정보를 훅의 상태와 동기화
   useEffect(() => {
-    // connectionStatus가 변경될 때마다 isConnected 업데이트
-    const isWsConnected = connectionStatus === 'connected' && wsList.length > 0;
-    setIsConnected(isWsConnected);
-    console.log(`🔌 WebSocket 연결 상태: ${connectionStatus}, 연결된 소켓: ${wsList.length}개, isConnected: ${isWsConnected}`);
-  }, [connectionStatus, wsList.length]);
-
-  // 이전 connectionId 추적을 위한 ref
-  const prevConnectionIdRef = useRef<string>('');
-
-  // connectionId 변경 시 비디오 스트리밍 갱신
-  useEffect(() => {
-    // 실제로 connectionId가 변경되었을 때만 처리
-    if (currentConnectionId &&
-      currentConnectionId !== prevConnectionIdRef.current &&
-      prevConnectionIdRef.current !== '') {
-      console.log('[LearnSession] connectionId 변경 감지:', prevConnectionIdRef.current, '->', currentConnectionId);
+    if (currentLessonSign) {
+      setCurrentSign(currentLessonSign);
+      setCurrentSignId(currentLessonSignId || '');
     }
-    // connectionId 업데이트
-    if (currentConnectionId) {
-      prevConnectionIdRef.current = currentConnectionId;
-    }
-  }, [currentConnectionId]);
-
-  // 현재 수어에 대한 ws url 출력
-  useEffect(() => {
-    if (currentSignId) {
-      console.log('[LearnSession] currentSignId:', currentSignId);
-      const wsUrl = lesson_mapper[currentSignId] || '';
-      setCurrentWsUrl(wsUrl);
-      console.log('[LearnSession] currentWsUrl:', wsUrl);
-
-      if (wsUrl) {
-        const connection = getConnectionByUrl(wsUrl);
-        if (connection) {
-          setCurrentConnectionId(connection.id);
-          setRetryAttempts(prev => ({ ...prev, wsConnection: 0 })); // 성공 시 재시도 카운터 리셋
-          console.log('[LearnSession] currentConnectionId:', connection.id);
-        } else {
-          console.warn(`[LearnSession] No connection found for targetUrl: ${wsUrl}, 재시도 시작`);
-          retryWsConnection(wsUrl);
-        }
-      } else {
-        console.warn('[LearnSession] currentSignId에 대한 WebSocket URL이 없음:', currentSignId);
-        // lesson_mapper에 해당 ID가 없으면 lesson_mapper 재시도
-        if (Object.keys(lesson_mapper).length === 0) {
-          retryLessonMapper();
-        }
-      }
-    }
-  }, [currentSignId, lesson_mapper, retryWsConnection, retryLessonMapper]);
-
-  // 소켓 메시지 수신 처리
-  useEffect(() => {
-    if (wsList && wsList.length > 0) {
-      // 각 소켓에 대해 핸들러 등록
-      const handlers: { ws: WebSocket; fn: (e: MessageEvent) => void }[] = [];
-      setMaxConfidence(0);
-
-      wsList.forEach(ws => {
-        const handleMessage = (event: MessageEvent) => {
-          try {
-            const msg = JSON.parse(event.data);
-            switch (msg.type) {
-              case 'classification_result': {
-
-                // 버퍼링 일시정지 중에 None 감지 시 버퍼링 재개
-                if (isBufferingPaused && msg.data && msg.data.prediction !== "None") {
-                  setDisplayConfidence("빠른 동작 감지");
-                  return;
-                } else if (isBufferingPaused && msg.data && msg.data.prediction === "None") {
-                  setIsBufferingPaused(false);
-                  return;
-                }
-
-
-                console.log('받은 분류 결과:', msg.data);
-                if (feedback && msg.data.prediction === "None") {
-                  setCurrentResult(msg.data);
-                  break;
-                }
-                const { prediction, confidence, probabilities } = msg.data;
-                const target = currentSign?.word;
-                let percent: number | undefined = undefined;
-                if (prediction === target) {
-                  percent = confidence * 100;
-                } else if (probabilities && target && probabilities[target] != null) {
-                  percent = probabilities[target] * 100;
-                }
-                if (percent != null) {
-                  setDisplayConfidence(`${percent.toFixed(1)}%`);
-                }
-                setCurrentResult(msg.data);
-                if (percent >= 80.0) {
-                  setFeedback("correct");
-                  studyListRef.current.push(currentSign.id);
-                  console.log("PASSED");
-                }
-                break;
-              }
-              default:
-                break;
-            }
-          } catch (e) {
-            console.error('WebSocket 메시지 파싱 오류:', e);
-          }
-        };
-        ws.addEventListener('message', handleMessage);
-        handlers.push({ ws, fn: handleMessage });
-      });
-
-      // 정리: 컴포넌트 언마운트 혹은 wsList 변경 시 리스너 해제
-      return () => {
-        handlers.forEach(({ ws, fn }) => {
-          ws.removeEventListener('message', fn);
-        });
-      };
-    }
-  }, [wsList, isBufferingPaused]);
-  //===============================================
-
-
+  }, [currentLessonSign, currentLessonSignId, setCurrentSign, setCurrentSignId]);
 
   //===============================================
   // 랜드마크 버퍼링 및 전송 처리
@@ -284,6 +103,7 @@ const LearnSession = () => {
       videoElement.playbackRate = isSlowMotion ? 0.5 : 1.0;
     }
   }, [isSlowMotion, videoSrc]);
+  
   // 랜드마크 감지 시 호출되는 콜백 (useCallback으로 먼저 정의)
   const handleLandmarksDetected = useCallback((landmarks: LandmarksData) => {
     // 녹화 중일 때만 버퍼에 추가
@@ -380,7 +200,7 @@ const LearnSession = () => {
         bufferIntervalRef.current = null;
       }
     };
-  }, [isRecording, isConnected, currentConnectionId, sendMessage, isBufferingPaused, currentResult]);
+  }, [isRecording, isConnected, currentConnectionId, sendMessage, isBufferingPaused, currentResult, setDisplayConfidence, setIsBufferingPaused]);
 
   useEffect(() => {
     setIsRecording(true);
@@ -449,13 +269,15 @@ const LearnSession = () => {
   // 애니메이션 재생 루틴
   const loadAnim = async () => {
     try {
-      const id = currentSign.id;
+      const id = currentLessonSign?.id;
+      if (!id) return;
+      
       console.log(id);
       const response = await API.get(`/anim/${id}`, {
         responseType: 'blob'
       });
 
-      const videoBlob = new Blob([response.data], { type: 'video/webm' });
+      const videoBlob = new Blob([response.data as BlobPart], { type: 'video/webm' });
       const videoUrl = URL.createObjectURL(videoBlob);
 
       if (videoSrc) {
@@ -470,7 +292,7 @@ const LearnSession = () => {
   // 수어 변경 시점마다 애니메이션 자동 변경
   useEffect(() => {
     loadAnim();
-  }, [currentSign]);
+  }, [currentLessonSign]);
 
   // 챕터 아이디를 통해 챕터 첫 준비
   useEffect(() => {
@@ -495,7 +317,6 @@ const LearnSession = () => {
   useEffect(() => {
     setCurrentSignIndex(0);
     setFeedback(null);
-    setCurrentResult(null); // 이전 분류 결과 초기화
 
     // 컴포넌트 언마운트 시 정리 작업 실시 
     return () => {
@@ -505,7 +326,7 @@ const LearnSession = () => {
         retryTimeoutRef.current = null;
       }
     };
-  }, []);
+  }, [setFeedback]);
 
   // 학습 세션 진입 시 /progress/lessons/events 호출 (mode: 'study')
   useEffect(() => {
@@ -576,7 +397,7 @@ const LearnSession = () => {
               autoStart={true}
               showControls={true}
               className="h-full"
-              currentSign={currentSign}
+              currentSign={currentLessonSign}
               currentResult={displayConfidence}
             />
             <Button
@@ -613,7 +434,7 @@ const LearnSession = () => {
           <div className="mt-8">
             <FeedbackDisplay
               feedback={feedback}
-              prediction={currentResult.prediction}
+              prediction={currentResult?.prediction}
               onComplete={feedback === 'correct' ? handleFeedbackComplete : undefined}
             />
           </div>
